backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.0
openai-whisper==20240930
torch==2.7.0
